browser/components/migration/tests/chrome/test_migration_wizard.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Basic tests for the Migration Wizard component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://browser/content/migration/migration-wizard.mjs" type="module"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script>
    "use strict";

    /**
     * Tests that the MigrationWizard:Init event is fired when the <migration-wizard>
     * is added to the DOM.
     */
    add_task(async function test_init_event() {
      let wiz = document.createElement("migration-wizard");
      let content = document.getElementById("content");
      let promise = new Promise(resolve => {
        content.addEventListener("MigrationWizard:Init", resolve, { once: true });
      })
      content.appendChild(wiz);
      await promise;
      ok(true, "Saw MigrationWizard:Init event.");
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content"></div>
<pre id="test"></pre>
</body>
</html>
